Align linking with requesting information
